Rename Users collection to Admin and update related references (#13)
