+++ src/app/example-double-admin-count/route.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, afterEach } from 'vitest'
 
-import { GET } from '@/app/example-double-user-count/route'
+import { GET } from '@/app/example-double-admin-count/route'
 
 import dotenv from 'dotenv'
 import { clearCollection, testPayloadObject } from '../../../tests/utils'
@@ -9,12 +9,12 @@ dotenv.config()
 
 describe('double user count', () => {
   afterEach(async () => {
-    await clearCollection(testPayloadObject, 'users')
+    await clearCollection(testPayloadObject, 'admins')
   })
 
   it('should double 1 user correctly', async () => {
     await testPayloadObject.create({
-      collection: 'users',
+      collection: 'admins',
       data: {
         email: 'rayzhao@gmail.com',
         password: '12132',
@@ -27,7 +27,7 @@ describe('double user count', () => {
 
   it('should double 2 user correctly', async () => {
     await testPayloadObject.create({
-      collection: 'users',
+      collection: 'admins',
       data: {
         email: 'rayzhao@gmail.com',
         password: '12132',
@@ -35,7 +35,7 @@ describe('double user count', () => {
     })
 
     await testPayloadObject.create({
-      collection: 'users',
+      collection: 'admins',
       data: {
         email: 'straight@gmail.com',
         password: '12132',
