+++ src/app/example-double-admin-count/route.ts
@@ -7,7 +7,7 @@ export const GET = async () => {
   })
 
   const data = await payload.find({
-    collection: 'users',
+    collection: 'admins',
   })
 
   return Response.json(data.docs.length * 2)
